Operator 클래스 추가 및 관련 메서드 테스트 코드 작성
